Moved files to follow new file structure, added windev batch for starting dev env on Windows, created basic UserDashboard component
