+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules*
 *venv*
 *__pycache__*
+build/
+dist/
