+++ README.md
@@ -3,10 +3,16 @@
 This will be a Django/Python API with a React frontend project for tracking bugs.
 
 ## Installing Flask
+
 1. Verify that Python is installed on your machine and Scripts are enabled (If on Windows, open cmd and type 'powershell Set-ExecutionPolicy RemoteSigned')
 2. In an open terminal, cd to the root of the project (Mothra/)
 3. type 'python -m venv venv'. This should create a new directory in Mothra/venv
 4. execute 'venv/Scripts/activate'
 5. execute 'pip install -r requirments.txt'
 6. cd into 'Mothra/backend-flask' and type 'flask run'
-The flask server should start running on 127.0.0.1:5000/
+   The flask server should start running on 127.0.0.1:5000/
+
+## Running Dev Environment
+
+In terminal, cd into 'Mothra/
+start windev
