+++ frontend/src/app/userDashboard/UserDashboard.module.css
@@ -1,0 +1,3 @@
+.userDashboard {
+  background-color: fuchsia;
+}
