+++ frontend/src/app/userDashboard/index.js
@@ -1,0 +1,1 @@
+export * from "./UserDashboard";
