+++ windev.bat
@@ -1,0 +1,4 @@
+cd backend-flask
+start flask run
+cd ../frontend
+npm start
